nvim: split test file
